refactor(repo): create `object`, `reference` modules
