+++ git-branchless-lib/src/git/mod.rs
@@ -3,7 +3,9 @@
 mod config;
 mod diff;
 mod index;
+mod object;
 mod oid;
+mod reference;
 mod repo;
 mod run;
 mod snapshot;
@@ -13,11 +15,14 @@ mod tree;
 pub use config::{Config, ConfigRead, ConfigValue, ConfigWrite};
 pub use diff::{process_diff_for_record, Diff};
 pub use index::{update_index, Index, IndexEntry, Stage, UpdateIndexCommand};
+pub use object::Commit;
 pub use oid::{MaybeZeroOid, NonZeroOid};
+pub use reference::{
+    Branch, BranchType, CategorizedReferenceName, Reference, ReferenceName, ReferenceTarget,
+};
 pub use repo::{
-    message_prettify, AmendFastOptions, Branch, BranchType, CategorizedReferenceName,
-    CherryPickFastError, CherryPickFastOptions, Commit, Error as RepoError, GitVersion, PatchId,
-    Reference, ReferenceName, ReferenceTarget, Repo, ResolvedReferenceInfo, Result as RepoResult,
+    message_prettify, AmendFastOptions, CherryPickFastError, CherryPickFastOptions,
+    Error as RepoError, GitVersion, PatchId, Repo, ResolvedReferenceInfo, Result as RepoResult,
     Time,
 };
 pub use run::{GitRunInfo, GitRunOpts, GitRunResult};
