+++ git-branchless-lib/src/git/repo.rs
@@ -9,37 +9,35 @@
 //! codebase.
 //! - To collect some different helper Git functions.
 
-use std::borrow::{Borrow, Cow};
+use std::borrow::Borrow;
 use std::collections::{HashMap, HashSet};
 use std::convert::{TryFrom, TryInto};
-use std::ffi::OsStr;
 use std::num::TryFromIntError;
 use std::ops::Add;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
-use std::string::FromUtf8Error;
 use std::time::{Duration, SystemTime};
 use std::{io, time};
 
-use bstr::{BString, ByteSlice, ByteVec};
+use bstr::ByteVec;
 use chrono::NaiveDateTime;
 use cursive::theme::BaseColor;
 use cursive::utils::markup::StyledString;
-use git2::{message_trailers_bytes, DiffOptions};
+use git2::DiffOptions;
 use itertools::Itertools;
 use thiserror::Error;
 use tracing::{instrument, warn};
 
 use crate::core::effects::{Effects, OperationType};
 use crate::core::eventlog::EventTransactionId;
-use crate::core::formatting::{Glyphs, StyledStringBuilder};
-use crate::core::node_descriptors::{
-    render_node_descriptors, CommitMessageDescriptor, CommitOidDescriptor, NodeObject, Redactor,
-};
+use crate::core::formatting::Glyphs;
 use crate::git::config::{Config, ConfigRead};
+use crate::git::object::Blob;
 use crate::git::oid::{make_non_zero_oid, MaybeZeroOid, NonZeroOid};
+use crate::git::reference::ReferenceNameError;
 use crate::git::run::GitRunInfo;
 use crate::git::tree::{dehydrate_tree, get_changed_paths_between_trees, hydrate_tree, Tree};
+use crate::git::{Branch, BranchType, Commit, Reference, ReferenceName};
 
 use super::index::{Index, IndexEntry};
 use super::snapshot::WorkingCopySnapshot;
@@ -1503,7 +1501,7 @@ impl Repo {
 
 /// The signature of a commit, identifying who it was made by and when it was made.
 pub struct Signature<'repo> {
-    inner: git2::Signature<'repo>,
+    pub(super) inner: git2::Signature<'repo>,
 }
 
 impl std::fmt::Debug for Signature<'_> {
@@ -1590,512 +1588,10 @@ pub struct PatchId {
     patch_id: git2::Oid,
 }
 
-/// Represents a commit object in the Git object database.
-#[derive(Clone, Debug)]
-pub struct Commit<'repo> {
-    inner: git2::Commit<'repo>,
-}
-
-impl<'repo> Commit<'repo> {
-    /// Get the object ID of the commit.
-    #[instrument]
-    pub fn get_oid(&self) -> NonZeroOid {
-        NonZeroOid {
-            inner: self.inner.id(),
-        }
-    }
-
-    /// Get the short object ID of the commit.
-    #[instrument]
-    pub fn get_short_oid(&self) -> Result<String> {
-        Ok(String::from_utf8_lossy(
-            &self
-                .inner
-                .clone()
-                .into_object()
-                .short_id()
-                .map_err(Error::Git)?,
-        )
-        .to_string())
-    }
-
-    /// Get the object IDs of the parents of this commit.
-    #[instrument]
-    pub fn get_parent_oids(&self) -> Vec<NonZeroOid> {
-        self.inner.parent_ids().map(make_non_zero_oid).collect()
-    }
-
-    /// Get the parent OID of this commit if there is exactly one parent, or
-    /// `None` otherwise.
-    #[instrument]
-    pub fn get_only_parent_oid(&self) -> Option<NonZeroOid> {
-        match self.get_parent_oids().as_slice() {
-            [] | [_, _, ..] => None,
-            [only_parent_oid] => Some(*only_parent_oid),
-        }
-    }
-
-    /// Get the number of parents of this commit.
-    #[instrument]
-    pub fn get_parent_count(&self) -> usize {
-        self.inner.parent_count()
-    }
-
-    /// Get the parent commits of this commit.
-    #[instrument]
-    pub fn get_parents(&self) -> Vec<Commit<'repo>> {
-        self.inner
-            .parents()
-            .map(|commit| Commit { inner: commit })
-            .collect()
-    }
-
-    /// Get the parent of this commit if there is exactly one parent, or `None`
-    /// otherwise.
-    #[instrument]
-    pub fn get_only_parent(&self) -> Option<Commit<'repo>> {
-        match self.get_parents().as_slice() {
-            [] | [_, _, ..] => None,
-            [only_parent] => Some(only_parent.clone()),
-        }
-    }
-
-    /// Get the commit time of this commit.
-    #[instrument]
-    pub fn get_time(&self) -> Time {
-        Time {
-            inner: self.inner.time(),
-        }
-    }
-
-    /// Get the summary (first line) of the commit message.
-    #[instrument]
-    pub fn get_summary(&self) -> Result<BString> {
-        match self.inner.summary_bytes() {
-            Some(summary) => Ok(BString::from(summary)),
-            None => Err(Error::DecodeUtf8 { item: "summary" }),
-        }
-    }
-
-    /// Get the commit message with some whitespace trimmed.
-    #[instrument]
-    pub fn get_message_pretty(&self) -> Result<BString> {
-        Ok(BString::from(self.inner.message_bytes()))
-    }
-
-    /// Get the commit message, without any whitespace trimmed.
-    #[instrument]
-    pub fn get_message_raw(&self) -> Result<BString> {
-        Ok(BString::from(self.inner.message_raw_bytes()))
-    }
-
-    /// Get the author of this commit.
-    #[instrument]
-    pub fn get_author(&self) -> Signature {
-        Signature {
-            inner: self.inner.author(),
-        }
-    }
-
-    /// Get the committer of this commit.
-    #[instrument]
-    pub fn get_committer(&self) -> Signature {
-        Signature {
-            inner: self.inner.committer(),
-        }
-    }
-
-    /// Get the `Tree` object associated with this commit.
-    #[instrument]
-    pub fn get_tree(&self) -> Result<Tree> {
-        let tree = self.inner.tree().map_err(|err| Error::FindTree {
-            source: err,
-            oid: self.inner.tree_id().into(),
-        })?;
-        Ok(Tree { inner: tree })
-    }
-
-    /// Get the "trailer" metadata from this commit's message. These are strings
-    /// like `Signed-off-by: foo` which appear at the end of the commit message.
-    #[instrument]
-    pub fn get_trailers(&self) -> Result<Vec<(String, String)>> {
-        let message = self.get_message_raw()?;
-        let message = message.to_str().map_err(|_| Error::DecodeUtf8 {
-            item: "raw messgae",
-        })?;
-        let mut result = Vec::new();
-        for (k, v) in message_trailers_bytes(message)
-            .map_err(Error::ReadMessageTrailer)?
-            .iter()
-        {
-            if let (Ok(k), Ok(v)) = (std::str::from_utf8(k), std::str::from_utf8(v)) {
-                result.push((k.to_owned(), v.to_owned()));
-            }
-        }
-        Ok(result)
-    }
-
-    /// Print a one-line description of this commit containing its OID and
-    /// summary.
-    #[instrument]
-    pub fn friendly_describe(&self, glyphs: &Glyphs) -> Result<StyledString> {
-        let description = render_node_descriptors(
-            glyphs,
-            &NodeObject::Commit {
-                commit: self.clone(),
-            },
-            &mut [
-                &mut CommitOidDescriptor::new(true).map_err(|err| Error::DescribeCommit {
-                    source: err,
-                    commit: self.get_oid(),
-                })?,
-                &mut CommitMessageDescriptor::new(&Redactor::Disabled).map_err(|err| {
-                    Error::DescribeCommit {
-                        source: err,
-                        commit: self.get_oid(),
-                    }
-                })?,
-            ],
-        )
-        .map_err(|err| Error::DescribeCommit {
-            source: err,
-            commit: self.get_oid(),
-        })?;
-        Ok(description)
-    }
-
-    /// Get a multi-line description of this commit containing information about
-    /// its OID, author, commit time, and message.
-    #[instrument]
-    pub fn friendly_preview(&self) -> Result<StyledString> {
-        let commit_time = self.get_time().to_naive_date_time();
-        let preview = StyledStringBuilder::from_lines(vec![
-            StyledStringBuilder::new()
-                .append_styled(
-                    format!("Commit:\t{}", self.get_oid()),
-                    BaseColor::Yellow.light(),
-                )
-                .build(),
-            StyledString::styled(
-                format!(
-                    "Author:\t{}",
-                    self.get_author()
-                        .friendly_describe()
-                        .unwrap_or_else(|| "".into())
-                ),
-                BaseColor::Magenta.light(),
-            ),
-            StyledString::styled(format!("Date:\t{}", commit_time), BaseColor::Green.light()),
-            StyledString::plain(textwrap::indent(
-                &self.get_message_pretty()?.to_str_lossy(),
-                "    ",
-            )),
-        ]);
-        Ok(preview)
-    }
-
-    /// Determine if the current commit is empty (has no changes compared to its
-    /// parent).
-    pub fn is_empty(&self) -> bool {
-        match self.get_parents().as_slice() {
-            [] => false,
-            [parent_commit] => self.inner.tree_id() == parent_commit.inner.tree_id(),
-            _ => false,
-        }
-    }
-
-    /// Determine if this commit added, removed, or changed the entry at the
-    /// provided file path.
-    #[instrument]
-    pub fn contains_touched_path(&self, path: &Path) -> Result<Option<bool>> {
-        let parent = match self.get_only_parent() {
-            None => return Ok(None),
-            Some(parent) => parent,
-        };
-        let parent_tree = parent.get_tree()?;
-        let current_tree = self.get_tree()?;
-        let parent_oid = parent_tree
-            .get_oid_for_path(path)
-            .map_err(Error::ReadTreeEntry)?;
-        let current_oid = current_tree
-            .get_oid_for_path(path)
-            .map_err(Error::ReadTreeEntry)?;
-        match (parent_oid, current_oid) {
-            (None, None) => Ok(Some(false)),
-            (None, Some(_)) | (Some(_), None) => Ok(Some(true)),
-            (Some(parent_oid), Some(current_oid)) => Ok(Some(parent_oid != current_oid)),
-        }
-    }
-
-    /// Amend this existing commit.
-    /// Returns the OID of the resulting new commit.
-    #[instrument]
-    pub fn amend_commit(
-        &self,
-        update_ref: Option<&str>,
-        author: Option<&Signature>,
-        committer: Option<&Signature>,
-        message: Option<&str>,
-        tree: Option<&Tree>,
-    ) -> Result<NonZeroOid> {
-        let oid = self
-            .inner
-            .amend(
-                update_ref,
-                author.map(|author| &author.inner),
-                committer.map(|committer| &committer.inner),
-                None,
-                message,
-                tree.map(|tree| &tree.inner),
-            )
-            .map_err(Error::Amend)?;
-        Ok(make_non_zero_oid(oid))
-    }
-}
-
-pub struct Blob<'repo> {
-    inner: git2::Blob<'repo>,
-}
-
-impl<'repo> Blob<'repo> {
-    pub fn get_content(&self) -> &[u8] {
-        self.inner.content()
-    }
-}
-
-/// The target of a reference.
-#[derive(Debug, PartialEq, Eq)]
-pub enum ReferenceTarget<'a> {
-    /// The reference points directly to an object. This is the case for most
-    /// references, such as branches.
-    Direct {
-        /// The OID of the pointed-to object.
-        oid: MaybeZeroOid,
-    },
-
-    /// The reference points to another reference with the given name.
-    Symbolic {
-        /// The name of the pointed-to reference.
-        reference_name: Cow<'a, OsStr>,
-    },
-}
-
-#[derive(Debug, Error)]
-pub enum ReferenceNameError {
-    #[error("reference name was not valid UTF-8: {0}")]
-    InvalidUtf8(FromUtf8Error),
-}
-
-/// The name of a reference, like `refs/heads/master`.
-#[derive(Clone, Debug, PartialOrd, Ord, PartialEq, Eq, Hash)]
-pub struct ReferenceName(String);
-
-impl ReferenceName {
-    /// Create a reference name from the provided bytestring. Non-UTF-8 references are not supported.
-    #[instrument]
-    pub fn from_bytes(bytes: Vec<u8>) -> std::result::Result<ReferenceName, ReferenceNameError> {
-        let reference_name = String::from_utf8(bytes).map_err(ReferenceNameError::InvalidUtf8)?;
-        Ok(Self(reference_name))
-    }
-
-    /// View this reference name as a string. (This is a zero-cost conversion.)
-    #[instrument]
-    pub fn as_str(&self) -> &str {
-        let Self(reference_name) = self;
-        reference_name
-    }
-}
-
-impl From<&str> for ReferenceName {
-    fn from(s: &str) -> Self {
-        ReferenceName(s.to_owned())
-    }
-}
-
-impl From<String> for ReferenceName {
-    fn from(s: String) -> Self {
-        ReferenceName(s)
-    }
-}
-
-impl From<NonZeroOid> for ReferenceName {
-    fn from(oid: NonZeroOid) -> Self {
-        Self::from(oid.to_string())
-    }
-}
-
-impl From<MaybeZeroOid> for ReferenceName {
-    fn from(oid: MaybeZeroOid) -> Self {
-        Self::from(oid.to_string())
-    }
-}
-
-impl AsRef<str> for ReferenceName {
-    fn as_ref(&self) -> &str {
-        &self.0
-    }
-}
-
-/// Represents a reference to an object.
-pub struct Reference<'repo> {
-    inner: git2::Reference<'repo>,
-}
-
-impl std::fmt::Debug for Reference<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self.inner.name() {
-            Some(name) => write!(f, "<Reference name={:?}>", name),
-            None => write!(f, "<Reference name={:?}>", self.inner.name_bytes()),
-        }
-    }
-}
-
-impl<'repo> Reference<'repo> {
-    /// Determine if the given name is a valid name for a reference.
-    pub fn is_valid_name(name: &str) -> bool {
-        git2::Reference::is_valid_name(name)
-    }
-
-    /// Get the name of this reference.
-    #[instrument]
-    pub fn get_name(&self) -> Result<ReferenceName> {
-        let name = ReferenceName::from_bytes(self.inner.name_bytes().to_vec())?;
-        Ok(name)
-    }
-    /// Get the commit object pointed to by this reference. Returns `None` if
-    /// the object pointed to by the reference is a different kind of object.
-    #[instrument]
-    pub fn peel_to_commit(&self) -> Result<Option<Commit<'repo>>> {
-        let object = match self.inner.peel(git2::ObjectType::Commit) {
-            Ok(object) => object,
-            Err(err) if err.code() == git2::ErrorCode::NotFound => return Ok(None),
-            Err(err) => return Err(Error::ResolveReference(err)),
-        };
-        match object.into_commit() {
-            Ok(commit) => Ok(Some(Commit { inner: commit })),
-            Err(_) => Ok(None),
-        }
-    }
-
-    /// Delete the reference.
-    #[instrument]
-    pub fn delete(&mut self) -> Result<()> {
-        self.inner.delete().map_err(Error::DeleteReference)?;
-        Ok(())
-    }
-}
-
-/// Determine what kind of branch a reference is, given its name. The returned
-/// `suffix` value is converted to a `String` to be rendered to the screen, so
-/// it may have lost some information if the reference name had unusual
-/// characters.
-///
-/// FIXME: This abstraction seems uncomfortable and clunky to use; consider
-/// revising.
-#[derive(Debug)]
-pub enum CategorizedReferenceName<'a> {
-    /// The reference represents a local branch.
-    LocalBranch {
-        /// The full name of the reference.
-        name: &'a str,
-
-        /// The string `refs/heads/`.
-        prefix: &'static str,
-    },
-
-    /// The reference represents a remote branch.
-    RemoteBranch {
-        /// The full name of the reference.
-        name: &'a str,
-
-        /// The string `refs/remotes/`.
-        prefix: &'static str,
-    },
-
-    /// Some other kind of reference which isn't a branch at all.
-    OtherRef {
-        /// The full name of the reference.
-        name: &'a str,
-    },
-}
-
-impl<'a> CategorizedReferenceName<'a> {
-    /// Categorize the provided reference name.
-    pub fn new(name: &'a ReferenceName) -> Self {
-        let name = name.as_str();
-        if name.starts_with("refs/heads/") {
-            Self::LocalBranch {
-                name,
-                prefix: "refs/heads/",
-            }
-        } else if name.starts_with("refs/remotes/") {
-            Self::RemoteBranch {
-                name,
-                prefix: "refs/remotes/",
-            }
-        } else {
-            Self::OtherRef { name }
-        }
-    }
-
-    /// Remove the prefix from the reference name. May raise an error if the
-    /// result couldn't be encoded as an `String` (shouldn't happen).
-    #[instrument]
-    pub fn remove_prefix(&self) -> Result<String> {
-        let (name, prefix): (_, &'static str) = match self {
-            Self::LocalBranch { name, prefix } => (name, prefix),
-            Self::RemoteBranch { name, prefix } => (name, prefix),
-            Self::OtherRef { name } => (name, ""),
-        };
-        Ok(name.strip_prefix(prefix).unwrap_or(name).to_owned())
-    }
-
-    /// Render the full name of the reference, including its prefix, lossily as
-    /// a `String`.
-    pub fn render_full(&self) -> String {
-        let name = match self {
-            Self::LocalBranch { name, prefix: _ } => name,
-            Self::RemoteBranch { name, prefix: _ } => name,
-            Self::OtherRef { name } => name,
-        };
-        (*name).to_owned()
-    }
-
-    /// Render only the suffix of the reference name lossily as a `String`. The
-    /// caller will usually check the type of reference and add additional
-    /// information to the reference name.
-    pub fn render_suffix(&self) -> String {
-        let (name, prefix): (_, &'static str) = match self {
-            Self::LocalBranch { name, prefix } => (name, prefix),
-            Self::RemoteBranch { name, prefix } => (name, prefix),
-            Self::OtherRef { name } => (name, ""),
-        };
-        name.strip_prefix(prefix).unwrap_or(name).to_owned()
-    }
-
-    /// Render the reference name lossily, and prepend a helpful string like
-    /// `branch` to the description.
-    pub fn friendly_describe(&self) -> String {
-        let name = self.render_suffix();
-        match self {
-            CategorizedReferenceName::LocalBranch { .. } => {
-                format!("branch {}", name)
-            }
-            CategorizedReferenceName::RemoteBranch { .. } => {
-                format!("remote branch {}", name)
-            }
-            CategorizedReferenceName::OtherRef { .. } => format!("ref {}", name),
-        }
-    }
-}
-
-/// Re-export of [`git2::BranchType`]. This might change to be an opaque type later.
-pub type BranchType = git2::BranchType;
-
 /// A timestamp as used in a [`Signature`].
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Time {
-    inner: git2::Time,
+    pub(super) inner: git2::Time,
 }
 
 impl Time {
@@ -2112,114 +1608,6 @@ impl Time {
     /// Calculate the associated [`NaiveDateTime`].
     pub fn to_naive_date_time(&self) -> NaiveDateTime {
         NaiveDateTime::from_timestamp(self.inner.seconds(), 0)
-    }
-}
-
-/// Represents a Git branch.
-pub struct Branch<'repo> {
-    repo: &'repo Repo,
-    inner: git2::Branch<'repo>,
-}
-
-impl std::fmt::Debug for Branch<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "<Branch name={:?}>",
-            String::from_utf8_lossy(
-                self.inner
-                    .name_bytes()
-                    .unwrap_or(b"(could not get branch name)")
-            ),
-        )
-    }
-}
-
-impl<'repo> Branch<'repo> {
-    /// Get the OID pointed to by the branch. Returns `None` if the branch is
-    /// not a direct reference (which is unusual).
-    pub fn get_oid(&self) -> Result<Option<NonZeroOid>> {
-        Ok(self.inner.get().target().map(make_non_zero_oid))
-    }
-
-    /// Get the name of this branch, not including any `refs/heads/` prefix. To get the full
-    /// reference name of this branch, instead call `.into_reference().get_name()?`.
-    #[instrument]
-    pub fn get_name(&self) -> eyre::Result<&str> {
-        self.inner
-            .name()?
-            .ok_or_else(|| eyre::eyre!("Could not decode branch name"))
-    }
-
-    /// Get the full reference name of this branch, including the `refs/heads/` or `refs/remotes/`
-    /// prefix, as appropriate
-    #[instrument]
-    pub fn get_reference_name(&self) -> eyre::Result<ReferenceName> {
-        let reference_name = self
-            .inner
-            .get()
-            .name()
-            .ok_or_else(|| eyre::eyre!("Could not decode branch reference name"))?;
-        Ok(ReferenceName(reference_name.to_owned()))
-    }
-
-    /// If this branch tracks a remote ("upstream") branch, return that branch.
-    #[instrument]
-    pub fn get_upstream_branch(&self) -> Result<Option<Branch<'repo>>> {
-        match self.inner.upstream() {
-            Ok(upstream) => Ok(Some(Branch {
-                repo: self.repo,
-                inner: upstream,
-            })),
-            Err(err) if err.code() == git2::ErrorCode::NotFound => Ok(None),
-            Err(err) => {
-                let branch_name = self.inner.name_bytes().map_err(|_err| Error::DecodeUtf8 {
-                    item: "branch name",
-                })?;
-                Err(Error::FindUpstreamBranch {
-                    source: err,
-                    name: String::from_utf8_lossy(branch_name).into_owned(),
-                })
-            }
-        }
-    }
-
-    /// If this branch tracks a remote ("upstream") branch, return the OID of the commit which that
-    /// branch points to.
-    #[instrument]
-    pub fn get_upstream_branch_target(&self) -> eyre::Result<Option<NonZeroOid>> {
-        let upstream_branch = match self.get_upstream_branch()? {
-            Some(upstream_branch) => upstream_branch,
-            None => return Ok(None),
-        };
-        let target_oid = upstream_branch.get_oid()?;
-        Ok(target_oid)
-    }
-
-    /// Get the associated remote to push to for this branch. If there is no
-    /// associated remote, returns `None`. Note that this never reads the value
-    /// of `push.remoteDefault`.
-    #[instrument]
-    pub fn get_push_remote_name(&self) -> eyre::Result<Option<String>> {
-        let branch_name = self
-            .inner
-            .name()?
-            .ok_or_else(|| eyre::eyre!("Branch name was not UTF-8: {self:?}"))?;
-        let config = self.repo.get_readonly_config()?;
-        if let Some(remote_name) = config.get(format!("branch.{branch_name}.pushRemote"))? {
-            Ok(Some(remote_name))
-        } else if let Some(remote_name) = config.get(format!("branch.{branch_name}.remote"))? {
-            Ok(Some(remote_name))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Convert the branch into its underlying `Reference`.
-    pub fn into_reference(self) -> Reference<'repo> {
-        Reference {
-            inner: self.inner.into_reference(),
-        }
     }
 }
